+++ BeatRevolution/Interface.h
@@ -73,7 +73,7 @@ class Interface
   void display_high_scores();
   void get_high_scores(int index);
   void display_play_data();
-  void upload_score(uint16_t score); // upload score, probably want to take score and maybe song name or username as params
+  void upload_score(uint32_t score); // upload score, probably want to take score and maybe song name or username as params
   void clear_screens();
 
   public:
